Media control using keyboard keys...
Press the left key on the keyboard to rewind 5 seconds.

Press the right key on the keyboard to go forwards 5 seconds.

Press the space key on the keyboard to play and pause the song.

Press the up key on the keyboard to increase the system volume.

Press the down key on the keyboard to decrease the system volume